// prisma/generated/zod/inputTypeSchemas/PersonWhereInputSchema.ts
export const PersonWhereInputSchema: z.ZodType<Prisma.PersonWhereInput> = z.object({
  AND: z.union([ z.lazy(() => PersonWhereInputSchema),z.lazy(() => PersonWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => PersonWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => PersonWhereInputSchema),z.lazy(() => PersonWhereInputSchema).array() ]).optional(),
  id: z.union([ z.lazy(() => UuidFilterSchema),z.string() ]).optional(),
  name: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  pronouns: z.lazy(() => StringNullableListFilterSchema).optional(),
  bio: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  updatedAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  isActive: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  preferredContactMethod: z.union([ z.lazy(() => EnumContactMethodNullableFilterSchema),z.lazy(() => ContactMethodSchema) ]).optional().nullable(),
  volunteerQuestionId: z.union([ z.lazy(() => IntFilterSchema),z.number() ]).optional(),
  user: z.union([ z.lazy(() => UserNullableRelationFilterSchema),z.lazy(() => UserWhereInputSchema) ]).optional().nullable(),
  personalInfo: z.union([ z.lazy(() => PIINullableRelationFilterSchema),z.lazy(() => PIIWhereInputSchema) ]).optional().nullable(),
  roleTerms: z.lazy(() => RoleTermListRelationFilterSchema).optional(),
  valuesWritten: z.lazy(() => EthicVersionListRelationFilterSchema).optional(),
  valueRankings: z.lazy(() => EthicsToPersonsListRelationFilterSchema).optional(),
  valueResponses: z.lazy(() => EthicVersionsToPersonsListRelationFilterSchema).optional(),
  circleRoles: z.lazy(() => CirclesToRolesListRelationFilterSchema).optional(),
  volunteerResponses: z.lazy(() => VolunteerResponseListRelationFilterSchema).optional(),
  volunterResponsesImported: z.lazy(() => VolunteerResponseListRelationFilterSchema).optional(),
  onboarding: z.lazy(() => OnboardingStepsToPersonsListRelationFilterSchema).optional(),
  availability: z.lazy(() => AvailabilityListRelationFilterSchema).optional(),
  volunteerQuestionEdits: z.lazy(() => VolunteerQuestionEditsListRelationFilterSchema).optional(),
  roleEdits: z.lazy(() => RoleEditsListRelationFilterSchema).optional(),
  circleAimEdits: z.lazy(() => CircleAimEditsListRelationFilterSchema).optional()
}).strict();